Create timeout test in monkey patching and add to Makefile
